delete images when delete a post

+++ Controllers/PostsController.cs
@@ -160,8 +160,11 @@ namespace GrowthDiary.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Content")] Post post)
+        public async Task<IActionResult> Edit(int id, [Bind] Post post)
         {
+            _logger.LogInformation(post.Id.ToString());
+            _logger.LogInformation(post.Content);
+
             if (id != post.Id)
             {
                 return NotFound();
@@ -171,8 +174,17 @@ namespace GrowthDiary.Controllers
             {
                 try
                 {
-                    _context.Update(post);
-                    await _context.SaveChangesAsync();
+                    var query = from pst in _context.Post where pst.Id == id select pst;
+                    var p = await query.SingleAsync();
+                    if (p != null)
+                    {
+                        var timeZoneInfo = TimeZoneInfo.FindSystemTimeZoneById("Asia/Tokyo");
+                        p.LastModifiedTime = TimeZoneInfo.ConvertTimeFromUtc(DateTime.UtcNow, timeZoneInfo);
+                        p.Content = post.Content;
+                        await _context.SaveChangesAsync();
+                        return RedirectToAction("Details", new { id });
+                    }
+                    return NotFound();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -213,9 +225,21 @@ namespace GrowthDiary.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var post = await _context.Post.FindAsync(id);
-            _context.Post.Remove(post);
-            await _context.SaveChangesAsync();
+            //var post = await _context.Post.FindAsync(id);
+            var query = from p in _context.Post.Include(p => p.Images)
+                        where p.Id == id
+                        select p;
+            var post = await query.SingleOrDefaultAsync();
+            if (post != null)
+            {
+                foreach (var image in post.Images)
+                {
+                    var path = Path.Combine(_environment.WebRootPath, image.Url.Substring(1));
+                    System.IO.File.Delete(path);
+                }
+                _context.Post.Remove(post);
+                await _context.SaveChangesAsync();
+            }
             return RedirectToAction(nameof(Index));
         }
 
@@ -223,13 +247,13 @@ namespace GrowthDiary.Controllers
         {
             return _context.Post.Any(e => e.Id == id);
         }
-        
+
         private IQueryable<Post> SearchPosts(String search, bool ascending)
         {
             var posts = _context.Post.AsQueryable();
             if (!String.IsNullOrEmpty(search))
                 posts = posts.Where(e => e.Content.Contains(search));
-                
+
             if (ascending)
                 posts = posts.OrderBy(e => e.CreationTime);
             else
